Refactor: rename 'singleton' module to 'payment_interface' and move it from 'common' to 'core.payments'
